Now observing bases from reads and generating observation signatures
main function now takes in phylotree input and a bam file of reads and
generates a table of read signature (a string that describes the observations
made at variant sites) and a list of read ids that follow this pattern. The
reduction in the number of signatures vs reads is not as great as I had
thought. This still may be a win at high coverage though.