Added Phylotree method to build table of haplotypes and variants.
Phylotree.get_haplotypes() build a dictionary of haplogroup ids mapped
to a list of variant strings. Unlike the previous way of building this we
now collapse haplogroups with identical variant descriptions into a single
entry for both. Also, removed member var variant_pos list and renamed
mutate_pos to take its place. Instead of storing a list of positions and
a dictionary of alleles and counts, we just store the dictionary and sort
the keys when we need to a list of positions.