HapVarBaseMatrix now takes number of mutations into account.
When finding the probability of observing a read from a specific haplotype,
HapVarBaseMatrix now used the number of mutations that have occured historicallyaccording to Phytree to weight the posibility of observed a unexpected base.
For example, a site that have been hit by 10 mutations will have a 90% chance
of observing the expected base for that haplotype and a 10% chance to observe
a different base (divided by 3 for the 3 other bases).