Fixed a big error when observing bases from reads.
Found a problem after adding mutation aware read probability calcualating.
When checking if a mapped position in a read has a variant, we were checking if
the query base was present rather than the reference base. This is a huge
mistake that slipped through because a lot of query positions are variant as
reference positions in the control region. This might have a big affect on the
results.