Read assignment now considers difference between top 2 hits for each read.
Previously, we assigned reads to haplogroups based on the raw "post."
probability. This did not take into account that this measurement is weighted
by the estimated proportion of the haplogroup, so reads that could equally be
from two haplogroups would have a higher probability of coming from the one
with the higher contribution to the mixture. We now compare the two highest
probabilities and divide each by the estimated proportion of the haplogroup.
If the ratio of these two is less than a cutoff (by default 2.0 right now) then
we cannot assign the read to either haplogroup.