Working on streamlining EM process.
Prior to this, running EM has gotten a little convoluted with the options for
convergence points, maximum iterations and number of complete iterations to
average over. This commit adds a new function em_step() for running a single
iteration of an EM run. This will be called by a new run_em() that will run
single and multiple EM runs while minimize the number matrices created.