Big commit: Refactored to remove read_sigs table, new haplotype assembly
Modified mixemt, assemble.py and preprocess.py to stop passing around table
mapping read observation strings (read signatures or sub-haplotype strings)
to the read identifiers from the BAM file. Instead of passing the read
signature labels for the matrix and this dictionary, we now pass a list
containing an entry for every row in the matrix. This entry is a list of the
read identifiers for the "sub-haplotype" that represents them. This is much
easier now that we do not have to pass around an extra table. The functions
that used this read_sigs table have been modified to get the read IDs
straight from the read list. Tests in `test/assemble_test.py` have been
changed remove the use of the read_sigs table.

Also, the existing functions for assembling contributor haplotypes have been
modified to fit the new approach. Before, we assigned reads only based on
indexes and waited until the very end to map the indexes to the reads in the
bamfile, when we output the reads associated with each contributor. Now, we
build a table that contains the AlignedSegment objects from pysam which allows
us to inspect the bases in the reads after the initial assignment. This will
be used to extend assemblies by finding new variants reads that overlap them.

The old version of assign_reads renamed assign_read_indexes and a new
assign_reads now returns a dictionary mapping contributor ids to lists of
AlignedSegment objects for the reads that have been assigned.