Added a fix for re-reading base observation by position table.
Previously, I changed the process_reads() function to work on an iterable that
could be a list of pysam AlignedSegments or an AlignmentFile. However, if it is
an AlignmentFile we need to be sure that it is reset to the beginning of the
file. Added a call to reset() to do this before handing it off the
process_reads().