Format change for saving results. Corrected errors in writing statistics.
Previously, we saved read IDs when writing intermediate EM results using a
comma-separated string note reads that represent the same sub-haplotype
observation. This was not the way to do it since a ',' is an acceptable char
to have in a SAM format QNAME. This now has been replaced with tabs.