assign_reads_from_new_vars() uses new table structure.
Previously used a dict of hapN ids to sets of query_names, but that was
unnecessarily complicated. We now store a dict of query_names to sets of hapN
ids and just remove any entry with more than one possible contributor. This
removes the need for a separate function to filter for reads with single
contributors + tests for that function.