Anonymous haplogroups are now ignored by default, can be enabled by -A
Anonymous haplogroups can cause a little trouble because they are likely to
be very similar to haplogroups that do have names and our naming convention
for anonymous haplogroups is not standard and maybe confusing. To avoid this
the use of anonymous haplogroups in the EM step is now not the default
behavior, but this can be enabled by the -A option. Anonymous nodes are still
stored in the tree, but are not kept in the Phylotree hap_var table.

Also,
- Added -A option to mixemt
- Added 2 tests for this new functionality.