Changed the -n --var-count flag.
Changed the way this option works to make more sense. The problem is that
we previously added this option to deal with the cases where a highly
divergent haplotype had a lot of allelic dropout. This is common in our 95:5%
mixtures. The way we handled it previous, we required that half of the
defining variants were present, but this implies that observing 2/5 variants
and 20/50 are basically the same. We now have an option (-n) that overrides
this in the case the number of variants reaches a certain level (10 or 15).
Think of it like this:

        |       .            .
        |      .  Accept  .
obs.    |     .        .
number  |    .      .
variant |   .    .
        |  .  .
        | ..         Reject
        |.
        +-------------------------
        Total N available variants

With -f option:

        |       .
        |      .
obs.    |     . Accept
number  |    .
variant |   .    .................
        |  .  .
        | ..         Reject
        |.
        +-------------------------
        Total N available variants

Modified tests to use this new scheme.