Added new method get_ancestral()
Phylotree now stores a reference to the reference sequence that the tree
is based on. This is used by get_ancestral() to return the ancestral bases
for all sites not affected by a mutation in a lineage.

We can refactor some of the code now so we do not have to pass a phylotree
object and a reference sequence string around all of the time.