Fix for problem in using ancestral bases as evidence for haplogroup.
Previously, we only checked positions with defining mutations to see if a
haplogroup was present in the sample. This led to problems where the defining
mutations on a branch we actually back mutations to the ancestral base. We
did not account for the fact that these ancestral bases were likely introduced
by haplogroups at higher proportions and that they should only be used as
evidence if the previous haplogroups did not carry the ancestral base. To fix
this, we now keep track of ancestral bases (positions where no mutation is
inferred to have occurred) so that they can be ignore when checking for the
defining mutations for a haplogroup/contributor.

Added, modified 2 tests to match new behavior and added 2 new ones.