HapVarBaseMatrix now pre-computes site instability score. Big win.
We now compute the site-specific instability frequency score for
every variant position and store it a dictionary at initialization.
The values are used later to find the probability of observing a
read from a haplogroup. This change cuts the time needed to build
the initial EM input matrix in half.