write_variants() output overhauled to be more useful.
write_variants used to report a list of polymorphic positions from phylotree
and whether they were expected to be polymorphic or fixed in the sample, given
the haplogroups that were detected. This has been changed to be more useful in
interpretting the results of mixemt. write_variants now reports for each
position in the reference, the number of A,C,G,Ts, whether this position is
expected to be polymorphic, whether this position would be called a variant
position under our scheme (i.e. whether there exist more than 1 base that pass
the minimum variant count cutoff) and, finally, the variants that are known for
this position from our detected haplogroups. This should make it much easier to
see how the results are affected by what bases are found in the sample.