Changed save/load to also include the input matrix for EM.
In order to implement the second EM run to refine the contribution estimates,
we need to know the original input matrix for EM. We could rebuild it but it
would take a long time and not necessarily match. Instead, the input matrix
is now output with the -s flag, and read in using the -l flag. This is a
little annoying since it means that any previous results will not be usable,
unless I add in some backwards compatibility checks.