Fixed issue with diagnostic variant check and custom haplogroups
Ran into a problem with the steps that fill in ancestral bases after a
haplogroup has been detected and the variant bases it carries are removed from
consideration as evidence for another haplogroup. In the case that a variant
that defines a custom haplogroup that is not at a known variant position, we
previously tried to delete that position from the list of ancestral pos/bases.
We now check to see if this position is included in the list before removing
it now. We may need to revisit this later.